+++ crates/sotf-plugins/crates/sotf-plugin-hal-input/src/lib.rs
@@ -2,8 +2,6 @@ pub mod params;
 
 #[path = "lib/hal_input_plugin.rs"]
 mod hal_input_plugin;
-#[path = "lib/misc.rs"]
-mod misc;
 #[path = "lib/types.rs"]
 mod types;