+++ crates/sotf-plugins/crates/sotf-plugin-hal-input/src/lib/misc.rs
@@ -1,3 +1,0 @@
-pub(super) fn default_channels() -> usize {
-    2
-}